# application/api.yaml
openapi: 3.0.1
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
paths:
...
  '/auth/caslogin':
    get:
      summary: Provides a session cookie given a valid CAS Service ticket
      # The service ticket can be used by the server to get user data from the SSO server
      # Should probably be provided by spring-security or java-cas-client
      externalDocs:
        description: Apereo CAS protocol documentation
        url: 'https://apereo.github.io/cas/4.2.x/protocol/CAS-Protocol.html'
      security: [] # You shouldn't have to be logged in to login
      tags: ['auth']
      parameters:
      - name: ticket
        in: query
        required: true
        schema:
          type: string
          example: 'ST-12345678'
      responses:
        '200':
          description: Successfully authenticated
          headers:
            Set-Cookie:
              schema:
                type: string
                example: JSESSIONID=abcde12345;
        # TODO : What to do if auth fails? 302?
  '/auth/login':
    post:
      summary: Provides a session cookie given a username and password
      security: [] # You shouldn't have to be logged in to login
      tags: ['auth']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        '200':
          description: Successfully authenticated
          headers:
            Set-Cookie:
              schema:
                type: string
                example: JSESSIONID=abcde12345;
        '400':
          description: Authentication failed
servers:
- url: '//localhost:8080/'
components:
  schemas:
...
    Credentials:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
      - username
      - password
...
  securitySchemes:
    authCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
...