# application/generated/src/main/java/com/etsmtl/codecrusade/controllers/AuthApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (3.3.2).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.etsmtl.codecrusade.controllers;

import com.etsmtl.codecrusade.model.Credentials;
import io.swagger.annotations.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2018-11-12T18:01:20.447505100-05:00[America/New_York]")

@Validated
@Api(value = "auth", description = "the auth API")
public interface AuthApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    @ApiOperation(value = "Provides a session cookie given a valid CAS Service ticket", nickname = "authCasloginGet", notes = "", tags={ "auth", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successfully authenticated") })
    @RequestMapping(value = "/auth/caslogin",
        method = RequestMethod.GET)
    default ResponseEntity<Void> authCasloginGet(@NotNull @ApiParam(value = "", required = true) @Valid @RequestParam(value = "ticket", required = true) String ticket) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    @ApiOperation(value = "Provides a session cookie given a username and password", nickname = "authLoginPost", notes = "", tags={ "auth", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successfully authenticated"),
        @ApiResponse(code = 400, message = "Authentication failed") })
    @RequestMapping(value = "/auth/login",
        consumes = { "application/json" },
        method = RequestMethod.POST)
    default ResponseEntity<Void> authLoginPost(@ApiParam(value = "" ,required=true )  @Valid @RequestBody Credentials credentials) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}